completely move user management api from nextcloud to ldap
