+++ src/web/backend/nextcloud/api.py
@@ -6,10 +6,6 @@ from edap import ConstraintError, MultipleObjectsFound, ObjectDoesNotExist
 from backend.utils import EncoderWithBytes
 from backend.ldap.utils import EdapMixin
 
-from backend.ldap.serializers import edap_users_schema, edap_user_schema
-from backend.ldap.api_serializers import api_user_schema, api_users_schema, api_franchises_schema, \
-    api_divisions_schema, api_teams_schema
-from backend.ldap.models import LdapUser
 from .utils import create_group_folder, get_nextcloud
 
 blueprint = Blueprint('nextcloud_api', __name__, url_prefix='/api')
@@ -31,102 +27,6 @@ class NextCloudMixin:
             'message': nextcloud_response.meta.get('message', ''),
             'data': nextcloud_response.data
             })
-
-
-class UserListViewSet(NextCloudMixin, EdapMixin, MethodView):
-
-    def get(self):
-        """ List users """
-        res = self.edap.get_users()
-        data = edap_users_schema.load(res).data
-        return jsonify(api_users_schema.dump(data).data)
-
-    def post(self):
-        """ Create user """
-        username = request.json.get('uid')
-        password = request.json.get('password')
-        name = request.json.get('name')
-        surname = request.json.get('surname')
-        if not all([username, password, name, surname]):
-            return jsonify({'message': 'username, password, name, surname fields are required'}), 400
-
-        user = api_user_schema.load(request.json).data
-        try:
-            res = user.create(password=password)
-        except ConstraintError as e:
-            return jsonify({'message': "Failed to create user. {}".format(e)}), 400
-
-        return jsonify(res)
-
-
-class UserRetrieveViewSet(NextCloudMixin,
-                          EdapMixin,
-                          MethodView):
-    """ ViewSet for single user """
-    def get(self, username):
-        """ List users """
-        try:
-            user = edap_user_schema.load(self.edap.get_user(username)).data
-        except MultipleObjectsFound:
-            return jsonify({'message': 'More than 1 user found'}), 409
-        except ObjectDoesNotExist:
-            return jsonify({'message': 'User does not exist'}), 404
-        user_groups = self.edap.get_user_groups(username)
-        user = {
-            **api_user_schema.dump(user).data,
-            "groups": [group for group in user_groups],
-            "franchises": api_franchises_schema.dump(user.get_franchises()).data,
-            "divisions": api_divisions_schema.dump(user.get_divisions()).data,
-            "teams": api_teams_schema.dump(user.get_teams()).data
-        }
-        return jsonify(user)
-
-    def delete(self, username):
-        """ Delete user """
-        # TODO: switch to edap
-        res = self.nextcloud.delete_user(username)
-        return self.nxc_response(res)
-
-    def patch(self, username, action=None):
-        # TODO: switch to edap
-        if action is not None:
-            if action not in ['enable', 'disable']:
-                return jsonify({}), 404
-            res = self.nextcloud.enable_user(username) if action == 'enable' else self.nextcloud.disable_user(username)
-            return self.nxc_response(res)
-        param = request.json.get('param')
-        value = request.json.get('value')
-        if not all([param, value]):
-            return jsonify({}), 400
-        res = self.nextcloud.edit_user(username, param, value)
-        return self.nxc_response(res)
-
-
-class UserGroupViewSet(NextCloudMixin,
-                       EdapMixin,
-                       MethodView):
-
-    def post(self, username):
-        """ Add user to group """
-        group_fqdn = request.json.get('fqdn')
-        if not group_fqdn:
-            return jsonify({'message': 'fqdn is required parameter'}), 400
-        try:
-            self.edap.make_uid_member_of(username, group_fqdn)
-        except ConstraintError as e:
-            return jsonify({'message': str(e)}), 404
-        return jsonify({'message': 'Success'}), 200
-
-    def delete(self, username):
-        """ Remove user from group """
-        group_fqdn = request.json.get('fqdn')
-        if not group_fqdn:
-            return jsonify({'message': 'fqdn is a required parameter'})
-        try:
-            res = self.edap.remove_uid_member_of(username, group_fqdn)
-        except ConstraintError as e:
-            return jsonify({'message': f'Failed to delete. {e}'}), 400
-        return jsonify({'message': 'Success'}), 202
 
 
 class GroupListViewSet(NextCloudMixin,
@@ -265,17 +165,6 @@ class GroupFolderViewSet(EdapMixin, NextCloudMixin, MethodView):
 
         return jsonify({"message": "Group folder successfully created"}), 201
 
-
-user_list_view = UserListViewSet.as_view('users_api')
-blueprint.add_url_rule('/users/', view_func=user_list_view, methods=['GET', 'POST'])
-
-user_view = UserRetrieveViewSet.as_view('user_api')
-blueprint.add_url_rule('/users/<username>', view_func=user_view, methods=['GET', 'DELETE'])
-blueprint.add_url_rule('/users/<username>', view_func=user_view, methods=['PATCH'])
-blueprint.add_url_rule('/users/<username>/<action>', view_func=user_view, methods=['PATCH'])
-
-user_group_view = UserGroupViewSet.as_view('user_groups_api')
-blueprint.add_url_rule('/users/<username>/groups/', view_func=user_group_view, methods=['POST', 'DELETE'])
 
 group_list_view = GroupListViewSet.as_view('groups_api')
 blueprint.add_url_rule('/groups/', view_func=group_list_view, methods=["GET", "POST", "DELETE"])
