+++ src/web/frontend/src/common/nextcloud-api.service.js
@@ -1,30 +1,5 @@
 import ApiService from './api.service'
 
-export const NxcUsersService = {
-  get (slug, action) {
-    return ApiService.get('users', slug)
-  },
-
-  delete (slug) {
-    return ApiService.delete('users', slug)
-  },
-
-  post (params) {
-    return ApiService.post('users', params)
-  }
-}
-
-export const NxcUserGroupsService = {
-  post (username, groupFqdn, action = '') {
-    let resource = action ? `users/${username}/groups/subadmins` : `users/${username}/groups`
-    return ApiService.post(resource, {'fqdn': groupFqdn})
-  },
-
-  delete (username, fqdn) {
-    return ApiService.delete(`users/${username}/groups`, '', {fqdn: fqdn})
-  }
-}
-
 export const NxcGroupsService = {
   get (slug, action = '', params) {
     slug = action ? `${slug}/${action}` : slug
@@ -66,4 +41,4 @@ export const GroupFolderService = {
   }
 }
 
-export default { NxcUsersService, NxcGroupsService, NxcUserGroupsService }
+export default { NxcGroupsService }
